simple_social_media/requirements.txt
django
misaka
django-bootstrap3
django-braces
